+++ Lumina_Event_Horizon_v1.3.7_Real_Extracted/shaders/program/composite.glsl
@@ -31,11 +31,7 @@ vec3 northVec = normalize(gbufferModelView[2].xyz);
 #ifdef OVERWORLD
     vec3 lightVec = sunVec * ((timeAngle < 0.5325 || timeAngle > 0.9675) ? 1.0 : -1.0);
 #elif defined END
-    vec3 endLightDirection = mat3(gbufferModelView) * -cameraPosition;
-    vec3 endLightFallback = mat3(gbufferModelView) * normalize(EVENT_HORIZON_DIRECTION);
-    vec3 lightVec = dot(endLightDirection, endLightDirection) > 0.00000001
-                  ? normalize(endLightDirection)
-                  : normalize(endLightFallback);
+    vec3 lightVec = sunVec;
 #else
     vec3 lightVec = sunVec;
 #endif